feat: interactibles setup and room transitions (#179)
* added r_click_update
player able to interact with objects in the room

* begun adding doors and setting up interactibles in the test level for build showcase
small changes to interactibles to facilitate behavior creation

* Player can now interact with doors and move to another room

* gui code moved into functions for readability purposes

* fixed bug
where not checking if we had reached the next node in the path correctly. If x or z velocity where 0 it would return true when it had not reached the destination

* bug fix:
fixed an issue where if the player was at the exact position as the door it was trying to interact with it would move a weird location and interact with the door from far away instead.